requirements.txt
asgiref==3.9.0
attrs==25.3.0
colorama==0.4.6
coverage==7.9.2
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Markdown==3.8.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
